requirements-djwasabi.txt
pytest-cov==2.11.1
responses==0.12.1
